xpdeint: Replace regular expression-based user code parser with a lexer.
IMPORTANT: As of this commit, to use xpdeint you must have Pygments (http://pygments.org/) installed.
To do so is just a simple 'easy_install Pygments'.

Updating the C++ code parser to a full lexer based on one provided by Pygments allows us to much better
understand the user's code. Previously, IP and EX operators in comments (or string constants) would be
interpreted as actual uses of IP or EX operators and treated accordingly, but that would probably not be
what the user intended. As a concrete example consider the IP operator 'L' used with the following code:
dphi_dt = 0.0; // L[phi] + someOtherStuff;
where the user has commented out the normal evolution and replaced it with nothing. In this case, previous
versions of xpdeint (and all versions of xmds) would pull out the L operator and still apply it in the evolution.
This will cause the user (who for the sake of argument, we'll call Stuart) much confusion.

By using a lexer, xpdeint becomes smarter and understands that it doesn't need to care about the L[phi] expression
because it is hidden by a comment.

Other improvements made by this change include much better checking of the correct use of IP operators. xpdeint now
recognises many more cases where IP operators are used incorrectly, for example 'dy_dt = L[x];'. Also the parsing of 
integer-valued dimension indices is improved by this update.

The code that provides this lexing functionality is in CodeLexer.py


One other minor change made in this update is the ability to perform the delta-a integration step in alternate spaces.
Just add the 'fourier_space' attribute on the 'integration_vectors' tag (see examples/wigner_spectral.xmds) to specify
the space of the delta-a (temporal derivative) operator. This is important if you want to use a proper spectral method
for solving PDEs. For many problems, solving them in spectral space can be faster as there are fewer large components
involved. Some of the vibrating string examples have seen approximately a 30% speed increase by rewriting the evolution
to occur in fourier space instead of in real space.

Another change: The way boundary conditions were being enforced in some of the dirichlet boundary condition examples 
(vibstring_circle, vibstring_ellipse, etc) was incorrect. They shouldn't be applied with a filter as a separate part
of the integration, but instead should be applied during the integration step itself.

Version number bumped to 0.5 "Are you thinking what I'm thinking B1?"


git-svn-id: https://xmds.svn.sourceforge.net/svnroot/xmds/trunk/xpdeint@2117 a0a4e76c-ba24-0410-a01c-b269113d594b