xpdeint: Simplify much code that used overridable properties with a 'lazy property' data descriptor.
The new 'lazyproperty' data descriptor defined in Utilities.py is like 'property' (which creates data attributes
whose values are provided by a function), however the value can be overridden by simply writing to the attribute.
As a side-benefit, the result of the method invocation used to calculate the property is automatically cached.


git-svn-id: https://xmds.svn.sourceforge.net/svnroot/xmds/trunk/xpdeint@2125 a0a4e76c-ba24-0410-a01c-b269113d594b