xpdeint: Simplify much code that used overridable properties with a 'lazy property' data descriptor.
The new 'lazyproperty' data descriptor defined in Utilities.py is like 'property' (which creates data attributes
whose values are provided by a function), however the value can be overridden by simply writing to the attribute.
As a side-benefit, the result of the method invocation used to calculate the property is automatically cached.


git-svn-id: svn+ssh://svn.code.sf.net/p/xmds/code/trunk/xpdeint@2125 41b19578-b77a-4e3b-9682-c8570d237623