xpdeint: Prevent generated code from varying between runs of xpdeint + Code cleanups.
Previously, the generated source code could actually slightly change between executions of xpdeint
on the same script. These changes are only in the ordering of the initialisation/evaluation of some
vectors for which the ordering should not matter. But in the name of safety it would be better if
these variations didn't exist. So in this commit, the hash of a vector is now based on its name
(something which won't change between executions of xpdeint) instead of on its location in memory
(which can and will change between executions).


git-svn-id: svn+ssh://svn.code.sf.net/p/xmds/code/trunk/xpdeint@2131 41b19578-b77a-4e3b-9682-c8570d237623