xpdeint: Fix a bug with sub-sampling fourier space with the distributed-mpi driver.
I thought this worked, but it turns out it didn't. The testcase that was supposed to check for this was broken.
Now the run_tests script will print a warning if the expected results output contains a NaN when adding a testcase.
This should make this less likely to happen in the future.
Also modify the run_tests script to separately state when NaNs are found in the output when they shouldn't be there.

Robin: I don't think you sub-sample k-space, so you should be fine. But be aware.


git-svn-id: https://xmds.svn.sourceforge.net/svnroot/xmds/trunk/xpdeint@2137 a0a4e76c-ba24-0410-a01c-b269113d594b