xpdeint: Significant internal changes to how blocks of user code are handled.
Previously, any object that used user code strings (contents of CDATA blocks) would separately
maintain the list of vector dependencies that block needed, the field to loop over it with and what
space to do it in. All this is now maintained by a CodeBlock object. The advantage of this change
is that it localises all transformations of user code to the CodeBlock class. i.e. the CodeBlock class
is now responsible for dealing with integer-valued dimension component access (e.g. phi[j, k-3]).

This will make it easier to implement other changes like the 'D_' operators which will be the easy
way to use derivatives (as fast as EX operators, not as fast as IP though). D operators will enable
the user to write code like "D_x2_y2[phi] + D_x[psi]" and get what the expect without understanding
the ins and outs of EX operators. It will also make it easier to change the way integer-dimension
component access works to a simpler and clearer notation.

All tests pass.


git-svn-id: https://xmds.svn.sourceforge.net/svnroot/xmds/trunk/xpdeint@2143 a0a4e76c-ba24-0410-a01c-b269113d594b