xpdeint: Update examples given recent change to integer-valued dimensions.
Added a warning when using adaptive integrators with stochastics indicating that the
results will not converge to the correct answer except in certain circumstances.

Added a testcase which tests having the integration in fourier space. We had an example that did
this, but also makes a good testcase. And fix a bug that prevented this from working because
we didn't have a testcase covering this.

Fix a bug where an exception would be raised when parsing a code block that had a variable name at the
end of the text.

Indent exceptions and warnings when printing them.

Fix the XML schema, it was reporting validation errors on dimension tags incorrectly.

Make function operators work again.



git-svn-id: https://xmds.svn.sourceforge.net/svnroot/xmds/trunk/xpdeint@2148 a0a4e76c-ba24-0410-a01c-b269113d594b