xpdeint: Add in a new, faster noise source, and fix some bugs.
Added a new noise source for Gaussian random numbers that uses the Ziggurat method
[G. Marsaglia and W.W. Tsang, Journal of Statistical Software 5, Issue 8 (2000)].
In my tests, this noise source has lead to simulation speedups of between 4 and 10%
over the dsfmt nosies (and between 10-20% over the default POSIX noises). The speedup
being more significant for higher-dimensional simulations. It is advisable to turn on
the 'auto_vectorise' feature when using the new noises.

To use the new Gaussian noise source, change your noise kind to "gaussian-solirte".
The dsfmt version of this noise is "gaussian-dsfmt", and the posix version is "gaussian"
and "gaussian-posix".

Also fixed a bug where integrators would not be allocated correctly when the allocation was
moved for optimisation reasons. This affected fibre.xmds, and wasn't noticed as there wasn't
a testcase for a stochastic simulation using EX/IP operators. So fibre.xmds has been added
as a testcase.



git-svn-id: https://xmds.svn.sourceforge.net/svnroot/xmds/trunk/xpdeint@2149 a0a4e76c-ba24-0410-a01c-b269113d594b