xpdeint: Regression (r2143): Fix boundary conditions for cross-propagation.
Since r2143, boundary condition code for cross-propagation wasn't actually used.
This wasn't picked up by the testcase because the original initialisation of the cross-propagation
vectors set them to the same value that the boundary-condition code would set them to. The testcase
has been changed so that this is no longer the case, and now the testcase passes.



git-svn-id: svn+ssh://svn.code.sf.net/p/xmds/code/trunk/xpdeint@2152 41b19578-b77a-4e3b-9682-c8570d237623