xpdeint: Regression (r2143): Fix boundary conditions for cross-propagation.
Since r2143, boundary condition code for cross-propagation wasn't actually used.
This wasn't picked up by the testcase because the original initialisation of the cross-propagation
vectors set them to the same value that the boundary-condition code would set them to. The testcase
has been changed so that this is no longer the case, and now the testcase passes.



git-svn-id: https://xmds.svn.sourceforge.net/svnroot/xmds/trunk/xpdeint@2152 a0a4e76c-ba24-0410-a01c-b269113d594b