xpdeint: Add missing function in distributed-mpi fourier transform implementation.
Previously, when looping over distributed vectors the loop may have extended over parts
of the array that were not strictly defined in the given space (because the allocated memory
per node can be larger than the currently required memory). This could only cause problems
during adaptive integration, and worst-case scenario is that it may have caused a slight reduction
in the timestep. The cause of the problem is that one function in the distributed-mpi fourier
transform implementation wasn't implemented. This has now been corrected.

All test cases pass.


git-svn-id: https://xmds.svn.sourceforge.net/svnroot/xmds/trunk/xpdeint@2156 a0a4e76c-ba24-0410-a01c-b269113d594b