xpdeint: Prevent people omitting derivatives in their integration blocks.
Currently, if you specify a vector in the integration vectors for an integrate block
and don't specify the 'dblah_dt' variables, Bad Things(TM) can and will happen that
make little sense (like the code running slowly). Now, if a user tries this, an exception
will be raised listing the missing variable.


git-svn-id: svn+ssh://svn.code.sf.net/p/xmds/code/trunk/xpdeint@2158 41b19578-b77a-4e3b-9682-c8570d237623