xpdeint: Prevent people omitting derivatives in their integration blocks.
Currently, if you specify a vector in the integration vectors for an integrate block
and don't specify the 'dblah_dt' variables, Bad Things(TM) can and will happen that
make little sense (like the code running slowly). Now, if a user tries this, an exception
will be raised listing the missing variable.


git-svn-id: https://xmds.svn.sourceforge.net/svnroot/xmds/trunk/xpdeint@2158 a0a4e76c-ba24-0410-a01c-b269113d594b