xpdeint: Add support for the calculation of spatial correlation functions like g(1) and g(2).
I had intended to implement this, but thanks to prodding from Karen Kheruntsyan, I have
finally gotten around to it. Basically, this commit adds support for dimension 'aliases'.
These dimension aliases are normal dimensions, but are intended to be used for the calculation
of correlation functions. You create an alias by adding the 'aliases' attribute to the dimension
tag within the <transverse_dimensions> tag. For example,
<dimension name="x" lattice="64"  domain="(-5, 5)" aliases="xp" />
creates a dimension called 'x' and a dimension called 'xp'. By creating this second dimension in
this way, any variable with the dimension 'x' can now be accessed with 'xp' if the dimension 'xp'
is being looped over. For example, say you want to access 'phi' at 'xp', then using the nonlocal
dimension access syntax this is just:
phi(x: xp)
meaning 'phi(x = xp)'. This also works in fourier space where both 'phi(kx: kxp)' and 'phi(kx: -kxp)'
work as expected. Examples of using this to calculate g1 and g2 are in the new testcase 
'fibre_with_correlation_functions.xmds' and 'groundstate_gaussian.xmds'.

Note: when using aliases, omitting the 'dimensions' specifier on vectors, etc will cause the vector
to be defined *only* for the non-alias dimensions. This seems like the most sensible behaviour. However,
the alias dimensions still exist and a vector can be explicitly created with this dimension by simply
listing all dimensions required for that vector.



git-svn-id: svn+ssh://svn.code.sf.net/p/xmds/code/trunk/xpdeint@2162 41b19578-b77a-4e3b-9682-c8570d237623