xpdeint: Remove dependence on xmds-1 for most simulations.
libxmds was only ever needed for using XSIL files for initialisation of vectors.
Most simulations don't do this, so it is a little silly to require xmds-1 to be
installed to be able to use this vast majority of simulations.

Also, added MIN macro. We already had a MAX macro defined, but forgot a MIN macro.


git-svn-id: https://xmds.svn.sourceforge.net/svnroot/xmds/trunk/xpdeint@2163 a0a4e76c-ba24-0410-a01c-b269113d594b