xpdeint: Remove dependence on xmds-1 for most simulations.
libxmds was only ever needed for using XSIL files for initialisation of vectors.
Most simulations don't do this, so it is a little silly to require xmds-1 to be
installed to be able to use this vast majority of simulations.

Also, added MIN macro. We already had a MAX macro defined, but forgot a MIN macro.


git-svn-id: svn+ssh://svn.code.sf.net/p/xmds/code/trunk/xpdeint@2163 41b19578-b77a-4e3b-9682-c8570d237623