xpdeint: Matrix transforms!
Version bump to 0.6 "A Physicist is an atom's way of knowing about atoms."

General matrix transforms using Gaussian quadrature for high-accuracy numerical integration
has been implemented. In this commit, the following transforms are available:
* Bessel transforms (Hermite Transforms). Useful for solving problems with cylindrical symmetry.
* Spherical Bessel transforms. Useful for solving problems with spherical symmetry.
* Hermite-Gauss transforms. Useful for solving problems using the harmonic oscillator basis.

All matrix transforms make use of the standard matrix algebra package BLAS. You will need to have the
'cblas' library installed on your system to use these transforms. BLAS may come with the LAPACK package.
Mac OS X comes with an optimised BLAS package.

Note that dimensions using matrix transforms will *not* have their points evenly spaced. This is because
of the use of Gaussian Quadrature for spectrally-accurate transforms.

To make use of these transforms ensure the '<mmt />' feature is included and then simply specify the transform
required on the 'dimension' tag. For example:
<dimension name="r" lattice="32" domain="(0.0, 1.0)" transform="bessel" />
<dimension name="s" lattice="40" domain="(0.0, 1.5)" transform="spherical-bessel" order="3" />
<dimension name="x" lattice="100" length_scale="1.5e-3" transform="hermite-gauss" />

For the 'bessel' and 'spherical-bessel' transforms an optional 'order' parameter specifies the order of the
bessel functions. i.e. choosing order="1" for 'bessel' uses as a basis the Bessel functions J_1(x), corresponding
to the solution of a problem where the solution is of the form phi(r, theta) = R(r) exp(i m theta). For the spherical
bessel functions, the order corresponds to the angular quantum number l.

The advantage of using bases like 'bessel' and 'spherical-bessel' is that the basis functions are eigenstates of the
radial component of the Laplacian operator. Hence the boundary condition at the origin is that of analyticity. This means
that instead of writing the laplacian as 1/r dR/dr + d^2R/dr^2, it can be written as '-kr*kr' in Bessel-space. This
completely removes all problems with divergent terms at the origin and any problems with doubling the coordinate to be
able to solve the problem using a Fourier basis. For an example on using the Bessel basis to solve the wave equation
on a disk see the example bessel_transform2.xmds. For an example combining this with a Cosine transform to solve for the
3D groundstate of a cigar-shaped condensate with reflection symmetry on one axis and rotation symmetry about that axis, see 
the testcase fast/bessel_cosine_groundstate.xmds

For hermite-gauss transforms, the 'length_scale' attribute on the 'dimension' object is the harmonic oscillator length scale
sqrt(hbar/(M*omega)). Note that this basis is special as the range of points are selected automatically to optimally represent
the number of basis functions chosen. Not only will the separation of the points change as the number of basis functions changes,
but so will the range. Note that the 'transformed' version of a dimension 'x' will be 'nx' where 'nx' will correspond to the
harmonic oscillator quantum number. For an example using this basis to solve for the groundstate of a Gross-Pitaevskii equation
see the example hermitegauss_groundstate.xmds.

For function bases with definite parity like the Hermite-Gauss, the faster Parity Matrix Multiplication Transform is used for a 50%
reduction in computation complexity for the transforms.

To make use of any of the Matrix transforms, the Python package 'mpmath' must be installed (easy_install mpmath). Note that this
package is *not* requried if you only intend to use the Fourier basis. To make use of the Spherical Bessel function basis you
will also need the GNU Scientific Library (GSL) installed.

Note that Stochastics do not currently work with Matrix transforms as the noises currently assume that the variance is the same
across the grid. This is not true for grids that are not uniformly spaced. I will also need to talk with people more familiar with
the Hermite-Gauss basis to determine if the correct spatial variance for the noises is the inverse of the Gaussian quadrature weight.

Calculating the abscissas for the matrix transforms can be time consuming, particularly for the Hermite-Gauss basis. I have tried to
limit this by ensuring that the abscissas are cached between runs of xpdeint if the ~/.xmds directory exists. In the new year I'll talk
to people more familiar with the Hermite-Gauss basis to determine if there is a faster way to calculate the abscissas.

Testcases and examples have been added for the matrix transforms.

Other changes:
* Testsuite results are now saved in testsuite_results. This ensures that the results aren't thrown out between reboots.
* A caching mechanism has been added to save any intensive computations that need to be performed by xpdeint itself. These
  results are saved in ~/.xmds if that path exists (otherwise not saved).




git-svn-id: svn+ssh://svn.code.sf.net/p/xmds/code/trunk/xpdeint@2178 41b19578-b77a-4e3b-9682-c8570d237623