xpdeint: Update to how transforms work.
It is no longer necessary to specify the 'mmt' feature in order to be able to use Matrix Transforms.
Instead, all that is necessary is to state the transform required. This also means that it is no longer
necessary to have <fftw version="none" /> if you don't want FFTW to be used in a simulation. Simply omitting
the fftw tag will work provided there are no dimensions that would default to using a 'dft' transform.

Also, preprocessing and postprocessing for Matrix Transforms making use of parity are now vectorised
when the auto_vectorise feature is turned on.


git-svn-id: https://xmds.svn.sourceforge.net/svnroot/xmds/trunk/xpdeint@2185 a0a4e76c-ba24-0410-a01c-b269113d594b