xpdeint: Add support for saving simulation output in HDF5 format.
HDF5 is the Heirachichal Data Format v5, a data format designed by
the National Center for Supercomputing Applications (NSCA). This is
a self-describing binary data format designed for the storage and
transfer of scientific data. It's basically what XSIL (Extensible
Scientific Interchange Language) tried to be, except it succeeded.
HDF5 is designed to support MPI and saving very large datasets.

For more information see http://www.hdfgroup.org/HDF5/

Also, the format for the output of breakpoints can be set independently
of the output format for the entire simulation. For example you can have
binary or HDF5 breakpoints but ASCII output.

Note that the HDF5 support does not currently extend to distributed MPI
simulations or data input, but both are planned.

You will need the HDF5 libraries installed to be able to save in this format.
Note that both Matlab and Mathematica support importing data from HDF5 files.

To use HDF5 output, simply change the format attribute on your breakpoint or
output elements to "hdf5". See vibstring.xmds for an example.

The testsuites and the XSILFile class do not currently support HDF5 data.


git-svn-id: https://xmds.svn.sourceforge.net/svnroot/xmds/trunk/xpdeint@2196 a0a4e76c-ba24-0410-a01c-b269113d594b