xpdeint: Support loading data from HDF5 files. (http://www.hdfgroup.org/HDF5/)
Currently only the 'strict' geometry-matching mode is supported, but I think I was the only
person to use the 'loose' geometry-matching mode anyway. That support will come later.
('strict' and 'loose' geometry-matching modes are documented in the xmds-1 docs).
Once support for 'loose' geometry-matching mode is implemented, xpdeint will not depend on
xmds-1 for any features. Before this commit, loading data from the results of a previous simulation
required reading the .xsil file, and that required libxmds. As of this commit, such functionality
can be obtained by storing and loading data using the HDF5 format. To use the HDF5 format, you
will need libhdf5 installed, and for that see the HDF5 website (above).

Note that this commit brings full support for loading HDF5 files in distributed-mpi simulations.

To load data from a HDF5 file use the following syntax inside your <vector> element:
<initialisation kind="hdf5">
  <filename>some_filename.h5</filename>
</initialisation>

This will load data by default from the "/1" group (which is where data is written in HDF5-format breakpoints).
To specify a different group (if, for example you're loading data exported from another application), add a
'group' attribute on the 'filename' tag.

Minor changes:
* Corrected the 'io/xsilloading.xmds' example. It was missing a 'dimension' attribute
  on the output which lead to behaviour other than was intended.
* Added 'io/mpi_xsilloading.xmds' example to test loading binary XSIL files with distributed-MPI.
* As a result, discovered and corrected a bug where a 1D vector would not be fourier transformed
  when one of its dimensions was one of the distributed dimensions in a distributed-mpi script.
* Added HDF5 versions of 'io/xsilloading.xmds' and 'io/mpi_xsilloading.xmds'.
* Cleaned up testcases and examples to remove the old 'version' attribute on 'fftw'.


git-svn-id: svn+ssh://svn.code.sf.net/p/xmds/code/trunk/xpdeint@2208 41b19578-b77a-4e3b-9682-c8570d237623