xpdeint: Add support for non-square matrix transforms.
This is useful for things like Projected-GP models that require a greater number of points
in the spatial representation than in the spectral representation. To use this, add a 
'spectral_lattice' attribute to the <dimension> element and this will set the number of points
used in the spectral basis. For example, to consider the first 20 harmonic-oscillator eigenstates,
but to use 30 points in the spatial representation use the following:

<dimension name="x" lattice="30" spectral_lattice="20" length_scale="1.0" transform="hermite-gauss" />

See examples/hermitegauss_transform.xmds for an example of the usage.


git-svn-id: https://xmds.svn.sourceforge.net/svnroot/xmds/trunk/xpdeint@2209 a0a4e76c-ba24-0410-a01c-b269113d594b