xpdeint: Add support for non-square matrix transforms.
This is useful for things like Projected-GP models that require a greater number of points
in the spatial representation than in the spectral representation. To use this, add a 
'spectral_lattice' attribute to the <dimension> element and this will set the number of points
used in the spectral basis. For example, to consider the first 20 harmonic-oscillator eigenstates,
but to use 30 points in the spatial representation use the following:

<dimension name="x" lattice="30" spectral_lattice="20" length_scale="1.0" transform="hermite-gauss" />

See examples/hermitegauss_transform.xmds for an example of the usage.


git-svn-id: svn+ssh://svn.code.sf.net/p/xmds/code/trunk/xpdeint@2209 41b19578-b77a-4e3b-9682-c8570d237623