xpdeint: Make sure that vectors are initialised in the correct order.
Vectors can depend on other vectors for initialisation (e.g. the Thomas-Fermi
wavefunction could depend on the potential). Previously the order of initialisation
did not take this into account. Now vectors are initialised in the correct order
so that all dependencies are evaluated first. If this results in a loop, xpdeint
will report an error.

Added vectors/initialisation_order.xmds testcase to cover this.


git-svn-id: svn+ssh://svn.code.sf.net/p/xmds/code/trunk/xpdeint@2212 41b19578-b77a-4e3b-9682-c8570d237623