xpdeint: Important bug fix. If you use stochastics and mpi-multi-path this affects YOU!
Previously, if you used the mpi-multi-path driver, the mean values returned were
larger than their correct values by the number of processes you were using. This
was not caught by the testsuite because of a bug in how the results are checked.

Once again, if you use the mpi-multi-path driver, you should update to this release.
I believe this bug has been present since the mpi-multi-path driver was created.
I am astounded that no one noticed this.


git-svn-id: https://xmds.svn.sourceforge.net/svnroot/xmds/trunk/xpdeint@2215 a0a4e76c-ba24-0410-a01c-b269113d594b