xpdeint: Never be caught out by C-style integer division again!
xpdeint & xmds users unfamiliar with the C programming language frequently
make the mistake of assuming that the result of an expression like '1/2' will
be '0.5' as it is mathematically. Unfortunately, C returns an integer as both
'1' and '2' are integers, giving the result '0'. xpdeint will now recognise this
problem in code and generate an _error_ and refuse to compile the script until
the user has corrected the issue. The error message describes the problem and
discusses how to go about fixing it.


git-svn-id: svn+ssh://svn.code.sf.net/p/xmds/code/trunk/xpdeint@2221 41b19578-b77a-4e3b-9682-c8570d237623