xpdeint: Change 'double' type to 'real'.
THIS CHANGE AFFECTS YOU!

Instead of setting the type of a vector to be 'double', you will now set it to be 'real'. 'double' will
no longer be accepted. Also, you should get in the habit of using 'real' as the type of floating point
numbers instead of 'double' as used previously. For example instead of
  
  const double mass = 1e-3;

You should write instead:

  const real mass = 1e-3;

Finally, in my last commit I said that accessing the real part of a complex variable had changed from
'psi.re' or 'real(psi)' to 'psi.real()' or 'Re(psi)'. It's a bit silly to have two different names for
this, so instead it will all just be 'Re(psi)' or 'psi.Re()'. To access the imaginary part you should use
'Im(psi)' or 'psi.Im()'. This works for setting the real or imaginary components as well.

All testcases and examples have been updated to use the new syntax.



git-svn-id: https://xmds.svn.sourceforge.net/svnroot/xmds/trunk/xpdeint@2242 a0a4e76c-ba24-0410-a01c-b269113d594b