Optimising generated simulations.
The most common equations xpdeint is used to solve are imaginary-time mean-field groundstates and real-time evolution of the Gross-Pitaevskii equation or similar. If you're using an adaptive integrator and an IP operator (you should be using IP over EX if you can), a significant fraction of the time is usually spent calculating complex exponentials.
In the case of the imaginary-time evolution, a real exponential is all that is needed and the imaginary part doesn't need to be computed. In this case, write your IP operator as:
<operator kind="ip" type="real">
  stuff
</operator>
This will enable the above optimisation.

Similarly, in the case of evolution, the argument to the complex exponential is purely imaginary and a similar optimisation is possible. In this case, write your operator as:
<operator kind="ip" type="imaginary">
  stuff
</operator>

Note that in neither of these cases are changes required to the definition of the IP operators themselves.
Hence if your IP operator is of the form:
T = -kx*kx;
then you can set type="real" safely.
If it is of the form:
T = -i*kx*kx;
then you can set type="imaginary" safely.
Of course, if you have both real and imaginary components,
T = -i*kx*kx - damping;
then you can omit the 'type' attribute and the IP operator will default to (as it has always done) the 'complex' type.

git-svn-id: svn+ssh://svn.code.sf.net/p/xmds/code/trunk/xpdeint@2305 41b19578-b77a-4e3b-9682-c8570d237623