More reliable auto-configuration and improved compatibility with user-provided libraries.
When auto-configuring, the generated C programs weren't previously executed to check that, for example, a library wasn't missing that is only required at runtime.

Also, when configuring xpdeint with an additional library search path (using the --lib-path), the library paths weren't also added to the runtime search path. Previously this would be fixed by modifying the LD_LIBRARY_PATH environment variable. This is no longer necessary.
Note that this problem only exists on Linux and Solaris systems.

git-svn-id: https://xmds.svn.sourceforge.net/svnroot/xmds/trunk/xpdeint@2306 a0a4e76c-ba24-0410-a01c-b269113d594b