More reliable auto-configuration and improved compatibility with user-provided libraries.
When auto-configuring, the generated C programs weren't previously executed to check that, for example, a library wasn't missing that is only required at runtime.

Also, when configuring xpdeint with an additional library search path (using the --lib-path), the library paths weren't also added to the runtime search path. Previously this would be fixed by modifying the LD_LIBRARY_PATH environment variable. This is no longer necessary.
Note that this problem only exists on Linux and Solaris systems.

git-svn-id: svn+ssh://svn.code.sf.net/p/xmds/code/trunk/xpdeint@2306 41b19578-b77a-4e3b-9682-c8570d237623