Obscure bug fix.
If you're using multiple operators blocks in one integrator for the same field AND using noises, xpdeint didn't know which one to generate noises for.
This is now fixed in a limited way.
Until noise vectors are implemented, xpdeint can only handle this case if one of the operators blocks doesn't require noises.
It is extremely unlikely that this bug affects you.

git-svn-id: svn+ssh://svn.code.sf.net/p/xmds/code/trunk/xpdeint@2312 41b19578-b77a-4e3b-9682-c8570d237623