Distributed MPI for all!
You can now use the distributed-mpi driver with all simulations that have two or more dimensions if the first two have transforms.
So Bessel/DCT, 3D Hermite-Gauss, DFT/DCT simulations can be distributed.

The syntax hasn't changed.

If you are interested in using this feature it is suggested that you double-check that your simulations are faster with MPI. This may not be the case for some matrix transforms when using only two processors.
Although this feature has been tested, you are advised to check that results aren't changed by using this.

git-svn-id: svn+ssh://svn.code.sf.net/p/xmds/code/trunk/xpdeint@2330 41b19578-b77a-4e3b-9682-c8570d237623