Fourier transforms for the Hermite-Gauss basis!
If 'x' is the name of the dimension then 'x' is the position basis, 'nx' is the Hermite-Gauss basis and 'kx' is the Fourier-transformed position basis.
Note that to retain spectral accuracy, the FFT is *not* used.

See the example hermitegauss_transform.xmds or the testcase transforms/hermitegauss_fourier.xmds for details.

Fixed a bug when using single-precision with some matrix transforms.

git-svn-id: https://xmds.svn.sourceforge.net/svnroot/xmds/trunk/xpdeint@2343 a0a4e76c-ba24-0410-a01c-b269113d594b