Fourier transforms for the Hermite-Gauss basis!
If 'x' is the name of the dimension then 'x' is the position basis, 'nx' is the Hermite-Gauss basis and 'kx' is the Fourier-transformed position basis.
Note that to retain spectral accuracy, the FFT is *not* used.

See the example hermitegauss_transform.xmds or the testcase transforms/hermitegauss_fourier.xmds for details.

Fixed a bug when using single-precision with some matrix transforms.

git-svn-id: svn+ssh://svn.code.sf.net/p/xmds/code/trunk/xpdeint@2343 41b19578-b77a-4e3b-9682-c8570d237623