Reorder dimensions in example bessel_cosine_groundstate.xmds
There is approximately an order of magnitude difference in performance between having the dimensions in the order (z, r) and (r, z). This is because the matrix transform libraries like having the dimensions they work on first. In the future it would be good to get xpdeint to automatically reorder the dimensions, but for the moment our examples should reflect best practise as people use them to create real simulations.

git-svn-id: svn+ssh://svn.code.sf.net/p/xmds/code/trunk/xpdeint@2352 41b19578-b77a-4e3b-9682-c8570d237623