Better error-checking of user code.
Previously, xpdeint checked only some of a user's code for
the common mistake of dividing two integers and expecting
the result to be a real number. This check was not applied
everywhere. The largest exception being the 'globals'
element. xpdeint now checks all of a user's code for such
mistakes and reports them. It is also easier to add other
such checks.

git-svn-id: svn+ssh://svn.code.sf.net/p/xmds/code/trunk/xpdeint@2356 41b19578-b77a-4e3b-9682-c8570d237623