Better error-checking of user code.
Previously, xpdeint checked only some of a user's code for
the common mistake of dividing two integers and expecting
the result to be a real number. This check was not applied
everywhere. The largest exception being the 'globals'
element. xpdeint now checks all of a user's code for such
mistakes and reports them. It is also easier to add other
such checks.

git-svn-id: https://xmds.svn.sourceforge.net/svnroot/xmds/trunk/xpdeint@2356 a0a4e76c-ba24-0410-a01c-b269113d594b