Add support for calculating 4-field Hermite-Gauss integrals exactly.
To use the 4-field quadrature rule instead of the 2-field quadrature rule
use the dimension 'x_4f' instead of 'x'. This is useful for transforming
the nonlinear potential term into spectral space in PGPE-type simulations.

The syntax of this feature is subject to change.

The example NLProjection.xmds demonstrates how to use this quadrature rule.

git-svn-id: svn+ssh://svn.code.sf.net/p/xmds/code/trunk/xpdeint@2371 41b19578-b77a-4e3b-9682-c8570d237623