Add patch to fix compilation of fftw-3 on Mac OS X Snow Leopard.
The latest version of fftw-3 does not compile on Snow Leopard.
A patch that corrects this problem has been added in admin/fftw3-SnowLeopard.patch
This patch has been submitted to the authors of FFTW, but I haven't heard anything back yet.
Instructions on how to apply the patch are in the README file.

git-svn-id: svn+ssh://svn.code.sf.net/p/xmds/code/trunk/xpdeint@2381 41b19578-b77a-4e3b-9682-c8570d237623