Minor fix for adaptive integrators.
For the simple case of dx/dt = 0 with x(t=0) = 0, it is impossible
for an adaptive integrator to determine a sensible relative error
estimate.  Although the difference between the solutions is 0,
the magnitude of the solution is 0 so the error is NaN.
Numerical codes don't much like NaNs, and so everything pretty much
fails after that point.  The change here is to handle NaNs errors by
setting the error to zero, as the only case in which this problem
can occur is the problem given above.

git-svn-id: svn+ssh://svn.code.sf.net/p/xmds/code/trunk/xpdeint@2382 41b19578-b77a-4e3b-9682-c8570d237623