Loose geometry matching mode supported for HDF5.
Loose geometry matching mode allows you to load data that was generated on
a slightly different grid into a simulation.  The restriction is that
the grid spacing in the exported basis must be the same.
While this may sound like a heavy restriction as the spacing between points
must remain the same, creative use of the bases in which data is exported
and imported can enable many things to be achieved.

For example, to load data that was generated on a coarser grid into a
simulation using a finer grid, simply export the grid in the first
simulation in Fourier space (e.g. using the basis 'kx').  If the second
simulation has the same width of the computational domain, then the spacing
between kx values will be the same, allowing more or less points to be used
in the second simulation.

To use loose geometry matching mode, specify geometry_matching_mode="loose"
on the initialisation tag for vectors.  For example,
    <initialisation kind="hdf5" geometry_matching_mode="loose">
        <filename>hermitegauss.h5</filename>
    </initialisation>

Many tests added for both binary and HDF5 loose geometry matching mode.
Oh, this also works with MPI :-).

git-svn-id: svn+ssh://svn.code.sf.net/p/xmds/code/trunk/xpdeint@2441 41b19578-b77a-4e3b-9682-c8570d237623