Fix for using noise vectors in a computed vector from a fixed step integrator with error checking on.
One problem was that the ErrorCheck feature was using old code. Another problem is that the preflight system is sensitive to ordering. There has been some overzealous use of @lazy_propery by me which has caused problems in the past.

The preflight system needs rethinking.
A workaround has been implemented now by adding a 'post_preflight' phase to deal with some of these ordering sensitivities in preflight, but this is a hack.

git-svn-id: svn+ssh://svn.code.sf.net/p/xmds/code/trunk/xpdeint@2457 41b19578-b77a-4e3b-9682-c8570d237623