Fix bug introduced in r2327.
Logic for when overridden indices are needed was wrong. It was incorrectly updated when updating code during the spaces -> basis conversion.

This affected delta a operators (a.k.a. dx_dt = blah) if aliasing was used.
It may have also affected some other things, not quite sure what.

Add test case to catch this situation.

Unrelated speedup of reset operation in adaptive integration.

git-svn-id: svn+ssh://svn.code.sf.net/p/xmds/code/trunk/xpdeint@2460 41b19578-b77a-4e3b-9682-c8570d237623