Made the messages printed during normal operation more user-friendly.
Up until now, xmds2 would print out the compile line that it would use to compile the script. This was visual noise that is not understood by most users. As a result, they do not always notice warnings printed by xmds2 or compile errors.

Now xmds2 just says when it is generating the simulation, and when it is compiling. If the compile fails, it will clearly say so.

To see the compile line, either use the '-g' option (debug), or the '-n' option (which means don't compile).

git-svn-id: svn+ssh://svn.code.sf.net/p/xmds/code/trunk/xpdeint@2461 41b19578-b77a-4e3b-9682-c8570d237623