Dimensions now have a 'volume_prefactor' attribute.
The volume_prefactor allows you to specify a metric term that should be incorporated in all integrals over that dimension.
For example, if you're using the DCT transform to represent something which is symmetric about the origin, write your dimension tag as:
    <dimension name="x" lattice="128" domain="(0, 1)" transform="dct" volume_prefactor="2.0" />
And any integrals over space (or fourier space) will essentially be over the domain (-1, 1).

It is recommended that you use volume_prefactor="2.0 * M_PI" with the 'bessel' transform, and volume_prefactor="4.0 * M_PI" with the 'spherical-bessel' transform.

Note that setting a value for volume_prefactor *will* alter the variance of any noises, but in a good way.  For example, if you wanted to calculate the number of atoms in a vacuum using Truncated-Wigner, set your initial state to:
phi = eta;
as you normally would (where eta is a complex gaussian variable), you may calculate the density of atoms using the usual:
mod2(phi) - 0.5/dx

git-svn-id: svn+ssh://svn.code.sf.net/p/xmds/code/trunk/xpdeint@2528 41b19578-b77a-4e3b-9682-c8570d237623