Output format is now HDF5 by default.
Changed install docs to call HDF5 a required dependency.
Updated install docs to reflect that MPI support is in the latest stable version of FFTW (no need to install alpha/beta).

Updated tutorials to use HDF5 by default, and remove the format attribute where unnecessary.

Updated some examples to remove the hdf5 format attribute as it is now the default.

git-svn-id: svn+ssh://svn.code.sf.net/p/xmds/code/trunk/xpdeint@2540 41b19578-b77a-4e3b-9682-c8570d237623