Fixes: 1) volume_prefactors now work for uniform dimensions with noises 2) Poissonian noise fields over nonuniform transverse dimensions will now compile
Change (fix?) 
- Poissonian noise fields now have integer-valued (cast as reals) values, with a mean rate per point equal to:
  	     Static noises: the volume of the element times a density
	     Dynamic noises: the same, multiplied by the propagation stepsize
Apart from a bug, we almost used to have this behaviour, but we'd then multiply the noise by the inverse volume element, 
which had the result that all elements effectively had the same mean count rate, regardless of their volume.  Interestingly,
we do not have an example that can tell the difference caused by this change, as no-one uses Poissonian noise fields it seems.   


git-svn-id: svn+ssh://svn.code.sf.net/p/xmds/code/trunk/xpdeint@2574 41b19578-b77a-4e3b-9682-c8570d237623