Generate SIMD instructions when detecting optimisation flags.
On Mac OS X, if you install a newer gcc, it is possible to get into the situation where the compiler supports AVX instructions, but the assembler doesn't.  If the configure tests don't generate these instructions, then we will accept optimisation flags that will fail with any actual simulation.

This way, we generate the SIMD instructions and hopefully trigger the same failure that would occur with an actual simulation.  We will then automatically fall back to a set of optimisation flags that work.

git-svn-id: svn+ssh://svn.code.sf.net/p/xmds/code/trunk/xpdeint@2590 41b19578-b77a-4e3b-9682-c8570d237623