Make the XMDS2 XML schema in the docs link to the relevent Reference entries
Also add placeholder entries in the Reference section for elements that haven't
been documented yet. It seems that trying to list the elements in a hierachical
level in the Reference section doesn't work very well, as some elements can
appear repeatedly at different levels in the tree. Maybe just give each element
the same header level? That way they'll also appear in the table of contents.


git-svn-id: svn+ssh://svn.code.sf.net/p/xmds/code/trunk/xpdeint@2599 41b19578-b77a-4e3b-9682-c8570d237623