Add facility to append argument values to output filename
Adds an optional attribute append_args_to_output_filename = "yes | "no"
to the Arguments element. If set to yes, the output filename will
have the names of the arguments passed in to the executable and their
values appended to the output filename. Useful for when you're doing
parameter scans. 

Arguments are added to the filename in alphabetic order, so the name
of the output filename is insensitive to the order in which the
parameters were passed.    

Breakpoint output filenames are currently unaffected by this. I'll
update them to also have this facility if no obvious bugs surface
in the near term.


git-svn-id: svn+ssh://svn.code.sf.net/p/xmds/code/trunk/xpdeint@2617 41b19578-b77a-4e3b-9682-c8570d237623