Linux installer tweaks and associated doc updates
Currently the linux_installer file is treated as having a MIME
type that results in a web page full of shell script when you
click on the download link. Giving it an .sh suffix means when
a user clicks on the link, they gets a prompt to save it instead.

Also bump revision installed by the installer to get latest
static library bug fixes.


git-svn-id: svn+ssh://svn.code.sf.net/p/xmds/code/trunk/xpdeint@2645 41b19578-b77a-4e3b-9682-c8570d237623