Implement single precision noises with the DSFMT generator
The single precision deviates are generated 25%-50% faster, but
since most of the time is spent calculating the distributions,
that RN generation speed is largely irrelevant. It seems that
exponentials, sqrts and so on are also faster with single precision,
however, so the generation of the distributions turns out to
also be faster.

In artificial, noise-generation-rate-bound simulations, single-
precision runs about 20%-70% faster(*). In most physically realistic
simulations, noise generation is a minor part of the simulation time,
but as usual if you're memory-bandwidth bound, single precision
can help a lot. The final gain for realistic single-precision noise
simulations that I've tried is 0%-60% over double precision.

(*) Unless you're using GCC 4.3 and using jump noises. In which
case it's eight times slower. Don't ask.


git-svn-id: svn+ssh://svn.code.sf.net/p/xmds/code/trunk/xpdeint@2646 41b19578-b77a-4e3b-9682-c8570d237623